pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "hatchling"]
build-backend = "hatchling.build"

[project]
name = "wq"
classifiers = ["Private :: Do Not Upload"]
version = "0"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-dotenv",
    "bs4",
    "llama-cpp-python",
    "chromadb",
]


[project.optional-dependencies]
dev = [
    "pip-tools",
    "isort",
    "black",
    "ruff",
    "tox",
]


[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
